Split implementation across multiple modules for further development
