+++ pyTorchAutoForge/datasets/ImagesAugmentation.py
@@ -32,6 +32,12 @@ from torchvision import transforms
 from pyTorchAutoForge.utils.conversion_utils import torch_to_numpy, numpy_to_torch
 from pyTorchAutoForge.datasets.DataAugmentation import AugsBaseClass
 
+from pyTorchAutoForge.datasets.noise_models.IntensityAugs import RandomBinarizeImage, RandomConvertTextureToShadingLaw, RandomNoiseTexturePattern
+from pyTorchAutoForge.datasets.noise_models.NoiseErrorsAugs import RandomGaussianNoiseVariableSigma, PoissonShotNoise
+
+from pyTorchAutoForge.datasets.noise_models.GeometricAugs import BorderAwareRandomAffine
+
+
 # %% Type aliases
 ndArrayOrTensor: TypeAlias = np.ndarray | torch.Tensor
 
@@ -39,8 +45,6 @@ ndArrayOrTensor: TypeAlias = np.ndarray | torch.Tensor
 # TODO modify to be usable by AugmentationSequential? Inherint from _AugmentationBase. Search how to define custom augmentations in Kornia
 
 # DEVNOTE issue with random apply: 0 is not allowed, but 1 is not either because it implies the user MUST specify at LEAST 2 augs. Easy workaround: automatically add a dummy that acts as a placeholder to make it work with 1
-
-
 class PlaceholderAugmentation(IntensityAugmentationBase2D):
 
     def __init__(self):
@@ -54,591 +58,7 @@ class PlaceholderAugmentation(IntensityAugmentationBase2D):
 
         return input
 
-# %% Intensity augmentations
-
-
-class PoissonShotNoise(IntensityAugmentationBase2D):
-    """
-    Applies Poisson shot noise to a batch of images.
-
-    This module simulates photon shot noise, where the variance of the noise 
-    is proportional to the pixel intensity. The noise is applied to a random 
-    subset of the batch based on the specified probability.
-
-    Args:
-        nn (torch.nn.Module): Base class for all neural network modules.
-
-    Attributes:
-        probability (float): Probability of applying Poisson noise to each 
-            image in the batch.
-
-    Methods:
-        forward(imgs_array: torch.Tensor) -> torch.Tensor:
-            Applies Poisson shot noise to the input batch of images.
-
-    Example:
-        >>> noise = PoissonShotNoise(probability=0.5)
-        >>> noisy_images = noise(images)
-    """
-
-    def __init__(self, probability: float = 0.0):
-        super().__init__(p=probability)
-
-    def apply_transform(self,
-                        input: torch.Tensor,
-                        params: dict[str, torch.Tensor],
-                        flags: dict[str, Any],
-                        transform: torch.Tensor | None = None) -> torch.Tensor:
-        """
-        Applies Poisson shot noise to the input batch of images.
-
-        Args:
-            x (torch.Tensor | tuple[torch.Tensor]): Input images as a tensor or tuple of tensors.
-            labels (torch.Tensor | tuple[torch.Tensor] | None, optional): Optional labels associated with the images.
-
-        Returns:
-            torch.Tensor | tuple[torch.Tensor, ...]: Images with Poisson shot noise applied, or a tuple containing such images.
-
-        """
-        # Randomly sample a boolean mask to index batch size
-        B = input.shape[0]
-        device, dtype = input.device, input.dtype
-
-        # Pixel value is the variance of the Photon Shot Noise (higher where brighter).
-        # Therefore, the mean rate parameter mu is equal to the DN at the specific pixel.
-        photon_shot_noise = torch.poisson(input).to(device=device)
-
-        # Sum noise to the original images
-        return input + photon_shot_noise
-
-    def generate_parameters(self,
-                            shape: tuple[int, ...]) -> dict:
-        return {}  # Not needed for this kind of noise
-
-    def compute_transformation(self, input: torch.Tensor, params: dict, flags: dict) -> torch.Tensor:
-        return torch.empty(0)  # Not used in IntensityAugmentationBase2D
-
-
-class RandomGaussianNoiseVariableSigma(IntensityAugmentationBase2D):
-    """
-    Applies per-sample Gaussian noise with variable sigma.
-    This augmentation adds Gaussian noise to each sample in a batch, where the standard deviation (sigma) can be a scalar, a (min, max) tuple for random sampling, or a per-sample array/tensor. The noise is applied to each sample with a specified probability.
-
-    Args:
-        sigma_noise (float or tuple[float, float]): Standard deviation of the Gaussian noise. Can be a scalar
-            or a tuple specifying the (min, max) range for random sampling per sample.
-        gaussian_noise_aug_prob (float, optional): Probability of applying noise to each sample. Defaults to 0.5.
-
-    Methods:
-        forward(x, labels=None):
-            Applies Gaussian noise to the input tensor with variable sigma per sample.
-
-    Example:
-        >>> aug = RandomGaussianNoiseVariableSigma(sigma_noise=(0.1, 0.5), gaussian_noise_aug_prob=0.7)
-        >>> noisy_imgs = aug(images)
-    """
-
-    def __init__(self,
-                 sigma_noise: int | float | tuple[float | int, float | int],
-                 gaussian_noise_aug_prob: float = 0.5,
-                 keep_scalar_sigma_fixed: bool = False,
-                 enable_img_validation_mode: bool = True,
-                 validation_min_num_bright_pixels: int = 50,
-                 validation_pixel_threshold: float = 5.0):
-
-        # Init superclass
-        super().__init__(p=gaussian_noise_aug_prob)
-
-        # Store parameters
-        self.sigma_gaussian_noise_dn = sigma_noise
-        self.keep_scalar_sigma_fixed: bool = keep_scalar_sigma_fixed
-
-        self.enable_img_validation_mode: bool = enable_img_validation_mode
-        self.validation_min_num_bright_pixels: int = validation_min_num_bright_pixels
-
-        # Pixel intensity threshold to consider a pixel "bright"
-        self.validation_pixel_threshold: float | Tensor = validation_pixel_threshold
-
-        # Initialization checks on sigma values
-        if isinstance(self.sigma_gaussian_noise_dn, (tuple, list)):
-            if len(self.sigma_gaussian_noise_dn) != 2:
-                raise ValueError(
-                    f'Invalid sigma_noise tuple length. Expected 2 got {len(self.sigma_gaussian_noise_dn)}')
-            if self.sigma_gaussian_noise_dn[0] < 0 or self.sigma_gaussian_noise_dn[1] < 0:
-                raise ValueError(
-                    'Sigma noise values must be non-negative.')
-            if self.sigma_gaussian_noise_dn[0] >= self.sigma_gaussian_noise_dn[1]:
-                raise ValueError(
-                    'Invalid sigma noise range. Minimum must be less than maximum.')
-            # Cast to tuple to ensure immutability
-            self.sigma_gaussian_noise_dn = tuple(
-                self.sigma_gaussian_noise_dn)  # type:ignore
-
-        elif isinstance(self.sigma_gaussian_noise_dn, (int, float)):
-            if self.sigma_gaussian_noise_dn < 0:
-                raise ValueError('Sigma noise value must be non-negative.')
-            # Cast to float
-            self.sigma_gaussian_noise_dn = float(self.sigma_gaussian_noise_dn)
-
-        else:
-            raise TypeError(
-                f'Invalid sigma_noise value. Expected int, float or tuple[float | int, float | int] got {type(self.sigma_gaussian_noise_dn)}')
-
-    def apply_transform(self,
-                        input: torch.Tensor,
-                        params: dict[str, torch.Tensor],
-                        flags: dict[str, Any],
-                        transform: torch.Tensor | None = None) -> torch.Tensor:
-
-        # Expect BCHW
-        if input.ndim != 4:
-            raise ValueError(
-                f"Expected 4D BCHW, got shape {tuple(input.shape)}")
-
-        B, C, H, W = input.shape
-        device = input.device
-
-        # Determine sigma per sample
-        sigma_val = self.sigma_gaussian_noise_dn
-
-        if isinstance(sigma_val, (tuple, list)):
-            # Get min-max values
-            min_s, max_s = sigma_val
-
-            # Do uniform sampling per sample
-            sigma_array = (max_s - min_s) * \
-                torch.rand(B, device=device) + min_s
-
-        elif isinstance(sigma_val, (float, int)):
-
-            if self.keep_scalar_sigma_fixed:
-                # Use same assigned sigma as fixed value
-                sigma_array = torch.full((B,), float(sigma_val), device=device)
-
-            else:
-                # Randomize using interval [0, sigma_val]
-                sigma_array = (sigma_val * torch.rand(B, device=device))
-        else:
-            raise TypeError(
-                f'Invalid sigma_noise value. Expected int, float or tuple[float | int, float | int] got {type(sigma_val)}')
-
-        # If validation mode, check image content first. Set sigma to zero for those entries before applying
-        if self.enable_img_validation_mode:
-
-            # Check threshold validity (auto-correct based on maximum value in images)
-            max_img_value = input.abs().max()
-            flat_input = input.abs().reshape(B, -1)
-
-            if self.validation_pixel_threshold >= max_img_value:
-
-                print(f"{colorama.Fore.LIGHTYELLOW_EX}WARNING: validation_pixel_threshold ({self.validation_pixel_threshold}) is greater than or equal to the maximum image pixel value ({max_img_value}). Adjusting threshold it automatically to 0.05 * median of max per sample: {self.validation_pixel_threshold}.{colorama.Style.RESET_ALL}")
-
-                # Get max per sample
-                max_per_sample = flat_input.max(dim=1).values
-
-                # Compute as 0.05 of median of max image values
-                validation_pixel_threshold_ = 0.05 * max_per_sample
-            else:
-                # Use the configured scalar threshold for all samples
-                validation_pixel_threshold_ = torch.full((B,),
-                                                         float(
-                    self.validation_pixel_threshold),
-                    device=input.device,
-                    dtype=flat_input.dtype,
-                )
-
-            # Count number of bright pixels per image in batch
-            is_pixel_bright_count_mask = flat_input.gt(validation_pixel_threshold_.view(
-                B, 1)).sum(dim=1)  # Get count of bright pixels per image
-
-            # Determine valid images mask
-            # Get bool mask where >= min_bright pixels shapes as (B,)
-            is_valid_mask = is_pixel_bright_count_mask.ge(
-                self.validation_min_num_bright_pixels)
-
-            # Set sigma to zero for invalid images by multiplying with bool mask
-            sigma_array = sigma_array * is_valid_mask.to(sigma_array.dtype)
-
-        # Sample image noise and apply to input batch
-        sigma_array = sigma_array.view(B, 1, 1, 1)
-        noise = torch.randn_like(input) * sigma_array
-
-        return input + noise
-
-
-class RandomNoiseTexturePattern(IntensityAugmentationBase2D):
-    """
-    Randomly fills the masked region with structured noise. 
-    Expects input tensor of shape (B, C+1, H, W):
-      - first C channels: image
-      - last    1 channel: binary mask (0=keep, 1=replace)
-    Returns the same shape.
-    """
-
-    def __init__(self,
-                 noise_texture_aug_prob: float = 0.5,
-                 randomization_prob: float = 0.6,
-                 masking_quantile: float = 0.85) -> None:
-        super().__init__(p=noise_texture_aug_prob)
-
-        self.randomization_prob = randomization_prob
-        self.masking_quantile = masking_quantile
-
-    def apply_transform(self,
-                        input: torch.Tensor,
-                        params: dict[str, torch.Tensor],
-                        flags: dict[str, Any],
-                        transform: torch.Tensor | None = None) -> torch.Tensor:
-
-        img_batch = input[0]
-        B = input.shape[0]
-        device, dtype = img_batch.device, img_batch.dtype
-
-        # Compute masks to select target patches
-        brightness = img_batch.mean(dim=1, keepdim=True)  # B×1×H×W
-        brightness_thr = (brightness.view(
-            B, 1, -1)).quantile(self.masking_quantile, dim=2, keepdim=True)
-
-        img_mask = (brightness > brightness_thr).float()
-
-        # TODO implement noise patterns to apply to image patches
-
-        # Re-attach mask channel for downstream modules
-        return out_imgs
-
-
-class RandomBinarizeImage(IntensityAugmentationBase2D):
-    def __init__(self,
-                 aug_prob: float = 0.5,
-                 masking_quantile: float = 0.15) -> None:
-        super().__init__(p=aug_prob)
-        self.masking_quantile = masking_quantile
-
-    def apply_transform(self,
-                        input: torch.Tensor,
-                        params: dict[str, torch.Tensor],
-                        flags: dict[str, Any],
-                        transform: torch.Tensor | None = None) -> torch.Tensor:
-
-        img_batch = input[0]
-        B = input.shape[0]
-        device, dtype = img_batch.device, img_batch.dtype
-
-        # Compute threshold
-        avg_brightness = img_batch.mean(dim=1, keepdim=True)  # B×1×H×W
-        brightness_thr = (avg_brightness.view(
-            B, 1, -1)).quantile(self.masking_quantile, dim=2, keepdim=True)
-
-        # Get binarized image
-        img_batch_mask = (avg_brightness > brightness_thr).float()
-
-        return img_batch_mask
-
-
-class RandomConvertTextureToShadingLaw(IntensityAugmentationBase2D):
-    """
-    RandomConvertTextureToShadingLaw Random augmentation that replaces textured regions with smooth shading based on selected shading law.
-    
-    _extended_summary_
-
-    :param IntensityAugmentationBase2D: _description_
-    :type IntensityAugmentationBase2D: _type_
-    """
-    def __init__(self,
-                    aug_prob: float = 0.5):
-        super().__init__(p=aug_prob)
-
-    def apply_transform(self,
-                        input: torch.Tensor,
-                        params: dict[str, torch.Tensor],
-                        flags: dict[str, Any],
-                        transform: torch.Tensor | None = None) -> torch.Tensor:
-        pass
-    
-
-# %% Geometric augmentations
-class BorderAwareRandomAffine(GeometricAugmentationBase2D):
-    """
-    BorderAwareRandomAffine Class reimplementing Kornia's RandomAffine with border crossing constraints.
-
-    This augmentation applies random affine transformations to a batch of images while considering border crossing constraints. It detects bright pixel runs along image borders to classify crossing types and modifies the affine transformation parameters accordingly before computing the transformation. The rest of the implementation follows RandomAffine. Depending on the detected crossing type, it constrains translations and rotations to prevent unrealistic augmentations. Detection is performed based on bright pixel runs along image borders (tailored for typical space images with near black background).
-
-    :param GeometricAugmentationBase2D: Base class for 2D geometric augmentations in Kornia.
-    :type GeometricAugmentationBase2D: class
-    """
-
-    def __init__(self,
-                 base_random_affine: K.RandomAffine,
-                 num_pix_crossing_detect: int = 10,
-                 intensity_threshold_uint8: float = 7.0):
-        super().__init__(p=base_random_affine.p,
-                         same_on_batch=base_random_affine.same_on_batch,
-                         keepdim=base_random_affine.keepdim)
-
-        # Input validity check
-        assert isinstance(
-            base_random_affine, K.RandomAffine), "base_random_affine must be an instance of K.RandomAffine"
-
-        # Store base random affine
-        self.base_random_affine = base_random_affine
-
-        # Get generator and flags from RandomAffine base
-        self._param_generator = base_random_affine._param_generator
-        self.flags = dict(resample=base_random_affine.flags["resample"],
-                          padding_mode=base_random_affine.flags["padding_mode"],
-                          align_corners=base_random_affine.flags["align_corners"])
-
-        # Store attributes
-        self.num_pix_crossing_detect = num_pix_crossing_detect
-        self.intensity_threshold_uint8 = intensity_threshold_uint8
-
-        # Assert validity of parameters
-        if self.num_pix_crossing_detect <= 0:
-            raise ValueError(
-                "num_pix_crossing_detect must be a positive integer.")
-        if self.intensity_threshold_uint8 < 0:
-            raise ValueError(
-                "intensity_threshold_uint8 must be a non-negative float.")
-
-    def compute_transformation(self,
-                               input: torch.Tensor,
-                               params: dict[str, torch.Tensor],
-                               flags: dict[str, Any]) -> torch.Tensor:
-        """
-        compute_transformation Computes the affine transformation matrix for each sample in the batch, taking into account border crossing constraints. Detection is performed based on bright pixel runs along image borders (tailored for typical space images with near black background).
-        """
-
-        # Get transform parameters (per sample)
-        translations = params["translations"]
-        angle = params["angle"]
-        scale = params["scale"]
-        shear_x = params["shear_x"]
-        shear_y = params["shear_y"]
-
-        # Get previously computed crossing type for this batch if available
-        crossing_types = params.get("crossing_types", None)
-
-        # Else compute crossing_types mask
-        if crossing_types is None:
-            crossing_types, _, _ = BorderAwareRandomAffine.Detect_border_crossing(
-                img_batch=input,
-                num_pix_detect=self.num_pix_crossing_detect,
-                intensity_threshold_uint8=self.intensity_threshold_uint8)
-            params["crossing_types"] = crossing_types
-
-        # Determine masks for transformations
-        crossing_types = crossing_types.to(device=translations.device)
-
-        vertical_only = crossing_types == 1
-        horizontal_only = crossing_types == 2
-        blocked = crossing_types == 3
-
-        # If any crossing found, modify params accordingly
-        if vertical_only.any() or horizontal_only.any() or blocked.any():
-
-            # Clone to avoid modifying original params
-            translations = translations.clone()
-            angle = angle.clone()
-            scale = scale.clone()
-            shear_x = shear_x.clone()
-            shear_y = shear_y.clone()
-
-            # Apply vertical only translation constraint (null out X translation)
-            if vertical_only.any():
-                translations[vertical_only, 0] = 0
-                angle[vertical_only] = 0
-                shear_x[vertical_only] = 0
-                shear_y[vertical_only] = 0
-                scale[vertical_only] = 1
-
-            # Apply horizontal only translation constraint (null out Y translation)
-            if horizontal_only.any():
-                translations[horizontal_only, 1] = 0
-                angle[horizontal_only] = 0
-                shear_x[horizontal_only] = 0
-                shear_y[horizontal_only] = 0
-                scale[horizontal_only] = 1
-
-            # Apply blocked constraint (null out all transformations)
-            if blocked.any():
-                translations[blocked] = 0
-                angle[blocked] = 0
-                shear_x[blocked] = 0
-                shear_y[blocked] = 0
-                scale[blocked] = 1
-
-            # Re-assign modified params
-            params["translations"] = translations
-            params["angle"] = angle
-            params["scale"] = scale
-            params["shear_x"] = shear_x
-            params["shear_y"] = shear_y
-
-        return KG.get_affine_matrix2d(
-            torch.as_tensor(params["translations"],
-                            device=input.device, dtype=input.dtype),
-            torch.as_tensor(params["center"],
-                            device=input.device, dtype=input.dtype),
-            torch.as_tensor(params["scale"],
-                            device=input.device, dtype=input.dtype),
-            torch.as_tensor(params["angle"],
-                            device=input.device, dtype=input.dtype),
-            KG.deg2rad(torch.as_tensor(
-                params["shear_x"], device=input.device, dtype=input.dtype)),
-            KG.deg2rad(torch.as_tensor(
-                params["shear_y"], device=input.device, dtype=input.dtype)),
-        )
-
-    def apply_transform(self,
-                        input: torch.Tensor,
-                        params: dict[str, torch.Tensor],
-                        flags: dict[str, Any],
-                        transform: torch.Tensor | None = None) -> torch.Tensor:
-
-        _, _, height, width = input.shape
-
-        if not isinstance(transform, torch.Tensor):
-            raise TypeError(
-                f"Expected the `transform` be a Tensor. Got {type(transform)}.")
-
-        # Apply affine warp
-        return KG.warp_affine(input,
-                              transform[:, :2, :],
-                              (height, width),
-                              flags["resample"].name.lower(),
-                              align_corners=flags["align_corners"],
-                              padding_mode=flags["padding_mode"].name.lower(),
-                              )
-
-    def inverse_transform(self,
-                          input: torch.Tensor,
-                          flags: dict[str, Any],
-                          transform: torch.Tensor | None = None,
-                          size: tuple[int, int] | None = None) -> torch.Tensor:
-
-        if not isinstance(transform, torch.Tensor):
-            raise TypeError(
-                f"Expected the `transform` be a Tensor. Got {type(transform)}.")
-
-        return self.apply_transform(input,
-                                    params=self._params,
-                                    transform=torch.as_tensor(
-                                        transform,
-                                        device=input.device,
-                                        dtype=input.dtype),            flags=flags,
-                                    )
-
-    @staticmethod
-    def Make_constrained_random_affines(base: K.RandomAffine):
-
-        gen = base._param_generator
-        tx, ty = (0.0, 0.0) if gen.translate is None else [
-            float(v) for v in gen.translate]
-
-        common_opts = dict(resample=base.flags["resample"],
-                           padding_mode=base.flags["padding_mode"],
-                           same_on_batch=base.same_on_batch,
-                           align_corners=base.flags["align_corners"],
-                           p=base.p,
-                           keepdim=base.keepdim,
-                           )
-
-        # Build constrained random affines
-        vertical_only = K.RandomAffine(degrees=0.0,
-                                       translate=(0.0, ty),
-                                       scale=None,
-                                       shear=None,
-                                       **common_opts)
-
-        horizontal_only = K.RandomAffine(degrees=0.0,
-                                         translate=(tx, 0.0),
-                                         scale=None,
-                                         shear=None, **common_opts)
-
-        return base, vertical_only, horizontal_only
-
-    @staticmethod
-    def Detect_border_crossing(img_batch,
-                               num_pix_detect: int = 10,
-                               intensity_threshold_uint8: float = 7.0):
-        """
-        Detect bright pixel runs along image borders to classify crossing types.
-
-        Returns:
-            crossing_type (torch.Tensor): Long tensor (B,) where 0=full/none, 1=vertical-only, 2=horizontal-only, 3=both.
-            vertical_crossing (torch.Tensor): Bool tensor (B,) true when left/right borders have a run.
-            horizontal_crossing (torch.Tensor): Bool tensor (B,) true when top/bottom borders have a run.
-        """
-        if img_batch.ndim != 4:
-            raise ValueError(
-                f"Expected BCHW tensor, got shape {tuple(img_batch.shape)}")
-
-        # Create bright pixel mask based on threshold (assumes uint8 input)
-        binarized_img_mask = (img_batch > intensity_threshold_uint8).any(dim=1)
-
-        B, H, W = binarized_img_mask.shape
-        device = binarized_img_mask.device
-
-        horizontal_edges = torch.cat(
-            (binarized_img_mask[:, 0, :], binarized_img_mask[:, -1, :]), dim=0)
-        vertical_edges = torch.cat(
-            (binarized_img_mask[:, :, 0], binarized_img_mask[:, :, -1]), dim=0)
-
-        # Determine crossings using convolutional count
-        horizontal_crossing = BorderAwareRandomAffine.ConvCount_intensity_line1d(
-            horizontal_edges, num_pix_detect)
-        vertical_crossing = BorderAwareRandomAffine.ConvCount_intensity_line1d(
-            vertical_edges, num_pix_detect)
-
-        # Get batch-wise masks
-        horizontal_crossing = horizontal_crossing[:B] | horizontal_crossing[B:]
-        vertical_crossing = vertical_crossing[:B] | vertical_crossing[B:]
-
-        # Determine crossing type
-        crossing_type = torch.full(
-            (B, 1), 0, device=device, dtype=torch.int64)  # 0: all allowed
-        crossing_type[vertical_crossing & ~
-                      horizontal_crossing] = 1  # 1: vertical only
-
-        crossing_type[horizontal_crossing & ~
-                      vertical_crossing] = 2  # 2: horizontal only
-
-        crossing_type[vertical_crossing &
-                      horizontal_crossing] = 3  # 3: none allowed
-
-        return crossing_type.squeeze(-1), vertical_crossing, horizontal_crossing
-
-    @staticmethod
-    def ConvCount_intensity_line1d(line_bool_mask: torch.Tensor,
-                                   window_length: int,
-                                   device: torch.device | str | None = None) -> torch.Tensor:
-
-        # Get length of line
-        length = line_bool_mask.shape[-1]
-
-        if device is None:
-            device = line_bool_mask.device
-
-        if length < window_length:
-            return torch.zeros(line_bool_mask.shape[0],
-                               device=device,
-                               dtype=torch.bool)
-
-        # Define convolution kernel
-        kernel = torch.ones((1, 1, window_length),
-                            device=device,
-                            dtype=torch.float32)
-
-        # Run convolution across line
-        conv = F.conv1d(input=line_bool_mask.float().unsqueeze(1),
-                        weight=kernel)
-
-        # Check if any window is fully filled
-        return conv.amax(dim=-1) >= window_length
-
 # %% Augmentation helper configuration dataclass
-
-
 @dataclass
 class AugmentationConfig:
     # Input specification
@@ -754,7 +174,6 @@ class AugmentationConfig:
 
 # %% Augmentation helper class
 # TODO (PC) add capability to support custom augmentation module by appending it in the user-specified location ("append_custom_module_after = (module, <literal>)" that maps to a specified entry in the augs_ops list. The given module is then inserted into the list at the specified position)
-
 class ImageAugmentationsHelper(nn.Module):
     def __init__(self, augs_cfg: AugmentationConfig):
         super().__init__()
